+++ tests/fscripts/demo_array6.fs
@@ -1,0 +1,6 @@
+a = array_create_with_str("1,2,3,4,0", ",");
+
+print(array_min(a))
+print(array_max(a))
+print(array_avg(a))
+print(array_sum(a))