+++ docs/changes.md
@@ -5,7 +5,8 @@
   * 增加 [字符串扩展模块](fscript_str.md)
   * fscript math 增加 atan2/log/is\_prime/exp 等函数。
   * fscript ext 增加 trim\_left/trim\_right/ulen/totitle/char\_at
-  * 完善mem\_allocator\_std内存统计（感谢QQ276310465提供补丁）
+  * 完善 mem\_allocator\_std 内存统计（感谢 QQ276310465 提供补丁）
+  * fscript array 增加 min/max/avg/sum 等函数。
 
 2021/07/15
   * 完善 widget\_factory，允许覆盖之前注册的控件。